Define __CHIBICC_6800_DOUBLE_AS_FLOAT__
- math.h needs the mode to choose the sin -> sinf macros, but headers cannot see compiler flags
- -fdouble-as-float defines the macro
- -fno-double-as-float undefines the macro
- The old __use_float_for_double.h wrapper is removed

+++ include/stdlib.h
@@ -16,7 +16,7 @@ int atoi(const char *nptr);
 long atol(const char *nptr);
 
 // Convert string to int (base 10 or autodetect)
-int strtol(const char *nptr, char **endptr, int base);
+long strtol(const char *nptr, char **endptr, int base);
 
 // Convert string to unsigned long
 unsigned long strtoul(const char *nptr, char **endptr, int base);
